requirements.txt
flask
python-dotenv
Werkzeug
py_crypto_hd_wallet
